feat(config): move workspace settings into Config page
