+++ packages/app/src/app/pages/settings.tsx
@@ -1,9 +1,8 @@
-import { For, Match, Show, Switch, createEffect, createMemo, createSignal, onCleanup, onMount } from "solid-js";
+import { For, Match, Show, Switch, createEffect, createMemo, createSignal, onMount } from "solid-js";
 
 import { formatBytes, formatRelativeTime, isTauriRuntime } from "../utils";
 
 import Button from "../components/button";
-import TextInput from "../components/text-input";
 import { HardDrive, MessageCircle, PlugZap, RefreshCcw, Shield, Smartphone, X } from "lucide-solid";
 import type { OpencodeConnectStatus, ProviderListItem, SettingsTab, StartupPreference } from "../types";
 import { createOpenworkServerClient } from "../lib/openwork-server";
@@ -53,13 +52,10 @@ export type SettingsViewProps = {
   openProviderAuthModal: () => Promise<void>;
   openworkServerStatus: OpenworkServerStatus;
   openworkServerUrl: string;
-  openworkServerSettings: OpenworkServerSettings;
   openworkServerHostInfo: OpenworkServerInfo | null;
   openworkServerCapabilities: OpenworkServerCapabilities | null;
   openworkServerDiagnostics: OpenworkServerDiagnostics | null;
   openworkServerWorkspaceId: string | null;
-  clientConnected: boolean;
-  canReloadWorkspace: boolean;
   openworkAuditEntries: OpenworkAuditEntry[];
   openworkAuditStatus: "idle" | "loading" | "error";
   openworkAuditError: string | null;
@@ -67,12 +63,6 @@ export type SettingsViewProps = {
   engineInfo: EngineInfo | null;
   openwrkStatus: OpenwrkStatus | null;
   owpenbotInfo: OwpenbotInfo | null;
-  reloadWorkspaceEngine: () => Promise<void>;
-  reloadBusy: boolean;
-  reloadError: string | null;
-  updateOpenworkServerSettings: (next: OpenworkServerSettings) => void;
-  resetOpenworkServerSettings: () => void;
-  testOpenworkServerConnection: (next: OpenworkServerSettings) => Promise<boolean>;
   developerMode: boolean;
   toggleDeveloperMode: () => void;
   stopHost: () => void;
@@ -112,11 +102,6 @@ export type SettingsViewProps = {
   downloadUpdate: () => void;
   installUpdateAndRestart: () => void;
   anyActiveRuns: boolean;
-  workspaceAutoReloadAvailable: boolean;
-  workspaceAutoReloadEnabled: boolean;
-  setWorkspaceAutoReloadEnabled: (value: boolean) => void | Promise<void>;
-  workspaceAutoReloadResumeEnabled: boolean;
-  setWorkspaceAutoReloadResumeEnabled: (value: boolean) => void | Promise<void>;
   onResetStartupPreference: () => void;
   openResetModal: (mode: "onboarding" | "all") => void;
   resetModalBusy: boolean;
@@ -135,7 +120,7 @@ export type SettingsViewProps = {
 };
 
 // Owpenbot Settings Component
-function OwpenbotSettings(props: {
+export function OwpenbotSettings(props: {
   busy: boolean;
   openworkServerStatus: OpenworkServerStatus;
   openworkServerUrl: string;
@@ -1186,28 +1171,6 @@ export default function SettingsView(props: SettingsViewProps) {
     }
   };
 
-  const [openworkUrl, setOpenworkUrl] = createSignal("");
-  const [openworkToken, setOpenworkToken] = createSignal("");
-  const [openworkTokenVisible, setOpenworkTokenVisible] = createSignal(false);
-  const [openworkTestState, setOpenworkTestState] = createSignal<"idle" | "testing" | "success" | "error">("idle");
-  const [openworkTestMessage, setOpenworkTestMessage] = createSignal<string | null>(null);
-  const [clientTokenVisible, setClientTokenVisible] = createSignal(false);
-  const [hostTokenVisible, setHostTokenVisible] = createSignal(false);
-  const [copyingField, setCopyingField] = createSignal<string | null>(null);
-  let copyTimeout: number | undefined;
-
-  createEffect(() => {
-    setOpenworkUrl(props.openworkServerSettings.urlOverride ?? "");
-    setOpenworkToken(props.openworkServerSettings.token ?? "");
-  });
-
-  createEffect(() => {
-    openworkUrl();
-    openworkToken();
-    setOpenworkTestState("idle");
-    setOpenworkTestMessage(null);
-  });
-
   const openworkStatusLabel = createMemo(() => {
     switch (props.openworkServerStatus) {
       case "connected":
@@ -1229,18 +1192,6 @@ export default function SettingsView(props: SettingsViewProps) {
         return "bg-gray-4/60 text-gray-11 border-gray-7/50";
     }
   });
-
-  const reloadAvailabilityReason = createMemo(() => {
-    if (!props.clientConnected) return "Connect to this workspace to reload.";
-    if (!props.canReloadWorkspace) {
-      return "Reloading is only available for local workspaces or connected OpenWork servers.";
-    }
-    return null;
-  });
-
-  const reloadButtonLabel = createMemo(() => (props.reloadBusy ? "Reloading..." : "Reload engine"));
-  const reloadButtonTone = createMemo(() => (props.anyActiveRuns ? "danger" : "secondary"));
-  const reloadButtonDisabled = createMemo(() => props.reloadBusy || Boolean(reloadAvailabilityReason()));
 
   const engineStatusLabel = createMemo(() => {
     if (!isTauriRuntime()) return "Unavailable";
@@ -1370,10 +1321,6 @@ export default function SettingsView(props: SettingsViewProps) {
         return "Model";
       case "advanced":
         return "Advanced";
-      case "remote":
-        return "Remote";
-      case "messaging":
-        return "Messaging Bridge";
       case "debug":
         return "Debug";
       default:
@@ -1382,7 +1329,7 @@ export default function SettingsView(props: SettingsViewProps) {
   };
 
   const availableTabs = createMemo<SettingsTab[]>(() => {
-    const tabs: SettingsTab[] = ["general", "model", "messaging", "remote", "advanced"];
+    const tabs: SettingsTab[] = ["general", "model", "advanced"];
     if (props.developerMode) tabs.push("debug");
     return tabs;
   });
@@ -1479,57 +1426,6 @@ export default function SettingsView(props: SettingsViewProps) {
     if (!uptimeMs) return "—";
     return formatRelativeTime(Date.now() - uptimeMs);
   };
-
-  const buildOpenworkSettings = () => ({
-    ...props.openworkServerSettings,
-    urlOverride: openworkUrl().trim() || undefined,
-    token: openworkToken().trim() || undefined,
-  });
-
-  const hasOpenworkChanges = createMemo(() => {
-    const currentUrl = props.openworkServerSettings.urlOverride ?? "";
-    const currentToken = props.openworkServerSettings.token ?? "";
-    return openworkUrl().trim() !== currentUrl || openworkToken().trim() !== currentToken;
-  });
-
-  const hostInfo = createMemo(() => props.openworkServerHostInfo);
-  const hostStatusLabel = createMemo(() => {
-    if (!hostInfo()?.running) return "Offline";
-    return "Available";
-  });
-  const hostStatusStyle = createMemo(() => {
-    if (!hostInfo()?.running) return "bg-gray-4/60 text-gray-11 border-gray-7/50";
-    return "bg-green-7/10 text-green-11 border-green-7/20";
-  });
-  const hostConnectUrl = createMemo(() => {
-    const info = hostInfo();
-    return info?.connectUrl ?? info?.mdnsUrl ?? info?.lanUrl ?? info?.baseUrl ?? "";
-  });
-  const hostConnectUrlUsesMdns = createMemo(() => hostConnectUrl().includes(".local"));
-
-  const handleCopy = async (value: string, field: string) => {
-    if (!value) return;
-    try {
-      await navigator.clipboard.writeText(value);
-      setCopyingField(field);
-      if (copyTimeout !== undefined) {
-        window.clearTimeout(copyTimeout);
-      }
-      copyTimeout = window.setTimeout(() => {
-        setCopyingField(null);
-        copyTimeout = undefined;
-      }, 2000);
-    } catch {
-      // ignore
-    }
-  };
-
-  onCleanup(() => {
-    if (copyTimeout !== undefined) {
-      window.clearTimeout(copyTimeout);
-    }
-  });
-
 
   return (
     <section class="space-y-6">
@@ -2025,324 +1921,6 @@ export default function SettingsView(props: SettingsViewProps) {
                 Requires typing <span class="font-mono text-gray-11">RESET</span> and will restart the app.
               </div>
             </div>
-          </div>
-        </Match>
-
-        <Match when={activeTab() === "remote"}>
-          <div class="space-y-6">
-            <Show when={hostInfo()}>
-              <div class="space-y-4">
-                <div class="bg-gray-2/30 border border-gray-6/50 rounded-2xl p-5 space-y-4">
-                  <div class="flex flex-col gap-2 md:flex-row md:items-center md:justify-between">
-                    <div>
-                      <div class="text-sm font-medium text-gray-12">OpenWork server sharing</div>
-                      <div class="text-xs text-gray-10">
-                        Share these details with a trusted device. Keep the server on the same network for the fastest setup.
-                      </div>
-                    </div>
-                    <div class={`text-xs px-2 py-1 rounded-full border ${hostStatusStyle()}`}>
-                      {hostStatusLabel()}
-                    </div>
-                  </div>
-
-                  <div class="grid gap-3">
-                    <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                      <div class="min-w-0">
-                        <div class="text-xs font-medium text-gray-11">OpenWork Server URL</div>
-                        <div class="text-xs text-gray-7 font-mono truncate">
-                          {hostConnectUrl() || "Starting server…"}
-                        </div>
-                        <Show when={hostConnectUrl()}>
-                          <div class="text-[11px] text-gray-8 mt-1">
-                            {hostConnectUrlUsesMdns()
-                              ? ".local names are easier to remember but may not resolve on all networks."
-                              : "Use your local IP on the same Wi-Fi for the fastest connection."}
-                          </div>
-                        </Show>
-                      </div>
-                      <Button
-                        variant="outline"
-                        class="text-xs h-8 py-0 px-3 shrink-0"
-                        onClick={() => handleCopy(hostConnectUrl(), "host-url")}
-                        disabled={!hostConnectUrl()}
-                      >
-                        {copyingField() === "host-url" ? "Copied" : "Copy"}
-                      </Button>
-                    </div>
-
-                    <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                      <div class="min-w-0">
-                        <div class="text-xs font-medium text-gray-11">Access token</div>
-                        <div class="text-xs text-gray-7 font-mono truncate">
-                          {clientTokenVisible()
-                            ? hostInfo()?.clientToken || "—"
-                            : hostInfo()?.clientToken
-                              ? "••••••••••••"
-                              : "—"}
-                        </div>
-                        <div class="text-[11px] text-gray-8 mt-1">Use on phones or laptops connecting to this server.</div>
-                      </div>
-                      <div class="flex items-center gap-2 shrink-0">
-                        <Button
-                          variant="outline"
-                          class="text-xs h-8 py-0 px-3"
-                          onClick={() => setClientTokenVisible((prev) => !prev)}
-                          disabled={!hostInfo()?.clientToken}
-                        >
-                          {clientTokenVisible() ? "Hide" : "Show"}
-                        </Button>
-                        <Button
-                          variant="outline"
-                          class="text-xs h-8 py-0 px-3"
-                          onClick={() => handleCopy(hostInfo()?.clientToken ?? "", "client-token")}
-                          disabled={!hostInfo()?.clientToken}
-                        >
-                          {copyingField() === "client-token" ? "Copied" : "Copy"}
-                        </Button>
-                      </div>
-                    </div>
-
-                    <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                      <div class="min-w-0">
-                        <div class="text-xs font-medium text-gray-11">Server token</div>
-                        <div class="text-xs text-gray-7 font-mono truncate">
-                          {hostTokenVisible()
-                            ? hostInfo()?.hostToken || "—"
-                            : hostInfo()?.hostToken
-                              ? "••••••••••••"
-                              : "—"}
-                        </div>
-                        <div class="text-[11px] text-gray-8 mt-1">Keep private. Required for approval actions.</div>
-                      </div>
-                      <div class="flex items-center gap-2 shrink-0">
-                        <Button
-                          variant="outline"
-                          class="text-xs h-8 py-0 px-3"
-                          onClick={() => setHostTokenVisible((prev) => !prev)}
-                          disabled={!hostInfo()?.hostToken}
-                        >
-                          {hostTokenVisible() ? "Hide" : "Show"}
-                        </Button>
-                        <Button
-                          variant="outline"
-                          class="text-xs h-8 py-0 px-3"
-                          onClick={() => handleCopy(hostInfo()?.hostToken ?? "", "host-token")}
-                          disabled={!hostInfo()?.hostToken}
-                        >
-                          {copyingField() === "host-token" ? "Copied" : "Copy"}
-                        </Button>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-              </div>
-            </Show>
-
-            <div class="space-y-4">
-              <div class="bg-gray-2/30 border border-gray-6/50 rounded-2xl p-5 space-y-4">
-                <div class="flex flex-col gap-2 md:flex-row md:items-center md:justify-between">
-                  <div>
-                    <div class="text-sm font-medium text-gray-12">OpenWork server</div>
-                    <div class="text-xs text-gray-10">
-                      Connect to an OpenWork server. Use the URL and access token from your server admin.
-                    </div>
-                  </div>
-                  <div class={`text-xs px-2 py-1 rounded-full border ${openworkStatusStyle()}`}>
-                    {openworkStatusLabel()}
-                  </div>
-                </div>
-
-                <div class="grid gap-3">
-                  <TextInput
-                    label="OpenWork server URL"
-                    value={openworkUrl()}
-                    onInput={(event) => setOpenworkUrl(event.currentTarget.value)}
-                    placeholder="http://127.0.0.1:8787"
-                    hint="Use the URL shared by your OpenWork server."
-                    disabled={props.busy}
-                  />
-
-                  <label class="block">
-                    <div class="mb-1 text-xs font-medium text-gray-11">Access token</div>
-                    <div class="flex items-center gap-2">
-                      <input
-                        type={openworkTokenVisible() ? "text" : "password"}
-                        value={openworkToken()}
-                        onInput={(event) => setOpenworkToken(event.currentTarget.value)}
-                        placeholder="Paste your token"
-                        disabled={props.busy}
-                        class="w-full rounded-xl bg-gray-2/60 px-3 py-2 text-sm text-gray-12 placeholder:text-gray-10 shadow-[0_0_0_1px_rgba(255,255,255,0.08)] focus:outline-none focus:ring-2 focus:ring-gray-6/20"
-                      />
-                      <Button
-                        variant="outline"
-                        class="text-xs h-9 px-3 shrink-0"
-                        onClick={() => setOpenworkTokenVisible((prev) => !prev)}
-                        disabled={props.busy}
-                      >
-                        {openworkTokenVisible() ? "Hide" : "Show"}
-                      </Button>
-                    </div>
-                    <div class="mt-1 text-xs text-gray-10">Optional. Paste the access token to authenticate.</div>
-                  </label>
-                </div>
-
-                <div class="text-[11px] text-gray-7 font-mono truncate">
-                  Resolved server: {openworkUrl().trim() || "Not set"}
-                </div>
-
-                <div class="flex flex-wrap gap-2">
-                  <Button
-                    variant="secondary"
-                    onClick={async () => {
-                      if (openworkTestState() === "testing") return;
-                      const next = buildOpenworkSettings();
-                      props.updateOpenworkServerSettings(next);
-                      setOpenworkTestState("testing");
-                      setOpenworkTestMessage(null);
-                      try {
-                        const ok = await props.testOpenworkServerConnection(next);
-                        setOpenworkTestState(ok ? "success" : "error");
-                        setOpenworkTestMessage(
-                          ok
-                            ? "Connection successful."
-                            : "Connection failed. Check the host URL and token."
-                        );
-                      } catch (error) {
-                        const message = error instanceof Error ? error.message : "Connection failed.";
-                        setOpenworkTestState("error");
-                        setOpenworkTestMessage(message);
-                      }
-                    }}
-                    disabled={props.busy || openworkTestState() === "testing"}
-                  >
-                    {openworkTestState() === "testing" ? "Testing..." : "Test connection"}
-                  </Button>
-                  <Button
-                    variant="outline"
-                    onClick={() => props.updateOpenworkServerSettings(buildOpenworkSettings())}
-                    disabled={props.busy || !hasOpenworkChanges()}
-                  >
-                    Save
-                  </Button>
-                  <Button
-                    variant="ghost"
-                    onClick={props.resetOpenworkServerSettings}
-                    disabled={props.busy}
-                  >
-                    Reset
-                  </Button>
-                </div>
-
-                <Show when={openworkTestState() !== "idle"}>
-                  <div
-                    class={`text-xs ${
-                      openworkTestState() === "success"
-                        ? "text-green-11"
-                        : openworkTestState() === "error"
-                          ? "text-red-11"
-                          : "text-gray-9"
-                    }`}
-                    role="status"
-                    aria-live="polite"
-                  >
-                    {openworkTestState() === "testing"
-                      ? "Testing connection..."
-                      : openworkTestMessage() ?? "Connection status updated."}
-                  </div>
-                </Show>
-
-                <Show when={openworkStatusLabel() !== "Connected"}>
-                  <div class="text-xs text-gray-9">
-                    OpenWork server connection needed to sync skills, plugins, and commands.
-                  </div>
-                </Show>
-              </div>
-
-              <div class="bg-gray-2/30 border border-gray-6/50 rounded-2xl p-5 space-y-4">
-                <div>
-                  <div class="text-sm font-medium text-gray-12">Engine reload</div>
-                  <div class="text-xs text-gray-10">Restart the OpenCode server for this workspace.</div>
-                </div>
-
-                <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                  <div class="min-w-0 space-y-1">
-                    <div class="text-sm text-gray-12">Reload now</div>
-                    <div class="text-xs text-gray-7">Applies config updates and reconnects your session.</div>
-                    <Show when={props.anyActiveRuns}>
-                      <div class="text-[11px] text-amber-11">Reloading will stop active tasks.</div>
-                    </Show>
-                    <Show when={props.reloadError}>
-                      <div class="text-[11px] text-red-11">{props.reloadError}</div>
-                    </Show>
-                    <Show when={reloadAvailabilityReason()}>
-                      <div class="text-[11px] text-gray-9">{reloadAvailabilityReason()}</div>
-                    </Show>
-                  </div>
-                  <Button
-                    variant={reloadButtonTone()}
-                    class="text-xs h-8 py-0 px-3 shrink-0"
-                    onClick={props.reloadWorkspaceEngine}
-                    disabled={reloadButtonDisabled()}
-                  >
-                    <RefreshCcw size={14} class={props.reloadBusy ? "animate-spin" : ""} />
-                    {reloadButtonLabel()}
-                  </Button>
-                </div>
-
-                <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                  <div class="min-w-0 space-y-1">
-                    <div class="text-sm text-gray-12">Auto reload (local)</div>
-                    <div class="text-xs text-gray-7">
-                      Reload automatically after agents/skills/commands/config change (only when idle).
-                    </div>
-                    <Show when={!props.workspaceAutoReloadAvailable}>
-                      <div class="text-[11px] text-gray-9">Available for local workspaces in the desktop app.</div>
-                    </Show>
-                  </div>
-                  <Button
-                    variant="outline"
-                    class="text-xs h-8 py-0 px-3 shrink-0"
-                    onClick={() => props.setWorkspaceAutoReloadEnabled(!props.workspaceAutoReloadEnabled)}
-                    disabled={props.busy || !props.workspaceAutoReloadAvailable}
-                  >
-                    {props.workspaceAutoReloadEnabled ? "On" : "Off"}
-                  </Button>
-                </div>
-
-                <div class="flex items-center justify-between bg-gray-1 p-3 rounded-xl border border-gray-6 gap-3">
-                  <div class="min-w-0 space-y-1">
-                    <div class="text-sm text-gray-12">Resume sessions after auto reload</div>
-                    <div class="text-xs text-gray-7">
-                      If a reload was queued while tasks were running, send a resume message afterward.
-                    </div>
-                  </div>
-                  <Button
-                    variant="outline"
-                    class="text-xs h-8 py-0 px-3 shrink-0"
-                    onClick={() => props.setWorkspaceAutoReloadResumeEnabled(!props.workspaceAutoReloadResumeEnabled)}
-                    disabled={props.busy || !props.workspaceAutoReloadAvailable || !props.workspaceAutoReloadEnabled}
-                    title={props.workspaceAutoReloadEnabled ? "" : "Enable auto reload first"}
-                  >
-                    {props.workspaceAutoReloadResumeEnabled ? "On" : "Off"}
-                  </Button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </Match>
-
-        <Match when={activeTab() === "messaging"}>
-          <div class="space-y-6">
-            <OwpenbotSettings
-              busy={props.busy}
-              openworkServerStatus={props.openworkServerStatus}
-              openworkServerUrl={props.openworkServerUrl}
-              openworkServerSettings={props.openworkServerSettings}
-              openworkServerWorkspaceId={props.openworkServerWorkspaceId}
-              openworkServerHostInfo={props.openworkServerHostInfo}
-              developerMode={props.developerMode}
-            />
           </div>
         </Match>
